LPD-X Merge ModelPreFilterContributorsRegistry/ModelPreFilterContributorsRegistryImpl into a pojo
