+++ modules/apps/portal-search/portal-search/src/main/java/com/liferay/portal/search/internal/indexer/helper/PreFilterContributorHelperImpl.java
@@ -7,6 +7,8 @@ package com.liferay.portal.search.internal.indexer.helper;
 
 import com.liferay.osgi.service.tracker.collections.list.ServiceTrackerList;
 import com.liferay.osgi.service.tracker.collections.list.ServiceTrackerListFactory;
+import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMap;
+import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMapFactory;
 import com.liferay.portal.kernel.exception.SystemException;
 import com.liferay.portal.kernel.model.BaseModel;
 import com.liferay.portal.kernel.search.BooleanClauseOccur;
@@ -86,11 +88,28 @@ public class PreFilterContributorHelperImpl
 	protected void activate(BundleContext bundleContext) {
 		_serviceTrackerList = ServiceTrackerListFactory.open(
 			bundleContext, SearchPermissionFilterContributor.class);
+
+		_classNameServiceTrackerMap =
+			ServiceTrackerMapFactory.openMultiValueMap(
+				bundleContext, ModelPreFilterContributor.class,
+				"indexer.class.name");
+
+		_mandatoryServiceTrackerMap =
+			ServiceTrackerMapFactory.openMultiValueMap(
+				bundleContext, ModelPreFilterContributor.class,
+				"indexer.clauses.mandatory");
+
+		_modelPreFilterContributorsRegistry =
+			new ModelPreFilterContributorsRegistry(
+				_classNameServiceTrackerMap, _mandatoryServiceTrackerMap);
 	}
 
 	@Deactivate
 	protected void deactivate() {
 		_serviceTrackerList.close();
+
+		_classNameServiceTrackerMap.close();
+		_mandatoryServiceTrackerMap.close();
 	}
 
 	protected Collection<String> getStrings(
@@ -100,10 +119,6 @@ public class PreFilterContributorHelperImpl
 			SearchStringUtil.splitAndUnquote(
 				(String)searchContext.getAttribute(string)));
 	}
-
-	@Reference
-	protected ModelPreFilterContributorsRegistry
-		modelPreFilterContributorsRegistry;
 
 	@Reference
 	protected QueryPreFilterContributorsRegistry
@@ -144,7 +159,7 @@ public class PreFilterContributorHelperImpl
 		SearchContext searchContext) {
 
 		List<ModelPreFilterContributor> modelPreFilterContributors =
-			modelPreFilterContributorsRegistry.filterModelPreFilterContributor(
+			_modelPreFilterContributorsRegistry.filterModelPreFilterContributor(
 				modelSearchSettings.getClassName(),
 				getStrings(
 					"search.full.query.clause.contributors.excludes",
@@ -250,6 +265,12 @@ public class PreFilterContributorHelperImpl
 		return entryClassName;
 	}
 
+	private ServiceTrackerMap<String, List<ModelPreFilterContributor>>
+		_classNameServiceTrackerMap;
+	private ServiceTrackerMap<String, List<ModelPreFilterContributor>>
+		_mandatoryServiceTrackerMap;
+	private ModelPreFilterContributorsRegistry
+		_modelPreFilterContributorsRegistry;
 	private ServiceTrackerList<SearchPermissionFilterContributor>
 		_serviceTrackerList;
